Tests/LibWeb/Text/expected/error-stack-must-contain-full-url.txt

classic: Error
    at Error
    at blob:<origin>/[flaky-uuid-replaced]:3:36

module: Error
    at Error
    at blob:<origin>/[flaky-uuid-replaced]:3:36
    at <unknown>
